Fix D1 and DO support
