+++ packages/django-cf/django_cf/db/backends/do/storage.py
@@ -1,0 +1,8 @@
+storage = None
+
+def set_storage(db):
+    global storage
+    storage = db
+
+def get_storage():
+    return storage
