+++ packages/django-cf/templates/d1/pyproject.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = "This template provides a starting point for running a Django application on Cloudflare Workers, utilizing Cloudflare D1 for serverless SQL database."
 requires-python = ">=3.12"
 dependencies = [
+    "django-cf",
     "django==5.2.6",
     "tzdata"
 ]
