--- packages/django-cf/templates/d1/src/django_cf/middleware/__init__.py
@@ -1,1 +1,0 @@
-from .CloudflareAccessMiddleware import CloudflareAccessMiddleware
