+++ packages/django-cf/templates/durable-objects/pyproject.toml
@@ -4,8 +4,9 @@ version = "0.1.0"
 description = "This template provides a starting point for running a Django application on Cloudflare Workers, utilizing Cloudflare Durable Objects for stateful data persistence."
 requires-python = ">=3.12"
 dependencies = [
-    "django-cf==0.1.5",
-    "django==5.1.2",
+    "django-cf",
+    "django==5.2.6",
+    "tzdata"
 ]
 
 [dependency-groups]
